JWT's added

+++ api/productapi/glassesapi.go
@@ -1,7 +1,9 @@
 package productapi
 
 import (
+	"finalproject/api"
 	"finalproject/products"
+	database "finalproject/user_database"
 	"strconv"
 
 	"github.com/gin-gonic/gin"
@@ -10,42 +12,64 @@ import (
 // Add Glasses
 func CreateGlassesCategory(c *gin.Context) {
 
-	products.InsertGlasses()
+	if database.VerifyAdmin(api.MyToken) {
 
-	c.JSON(200, gin.H{
-		"message": "Glasses category created successfully",
-	})
+		products.InsertGlasses()
+
+		c.JSON(200, gin.H{
+			"message": "Glasses category created successfully",
+		})
+	} else {
+		c.JSON(200, gin.H{
+			"message": "You are not authorized to perform this action",
+		})
+	}
 }
 
 func AddGlasses(c *gin.Context) {
 
-	Brand := c.Params.ByName("brand")
-	Name := c.Params.ByName("name")
-	Price := c.Params.ByName("price")
-	Stock := c.Params.ByName("stock")
-	Size := c.Params.ByName("size")
-	Rating := c.Params.ByName("rating")
-	Gender := c.Params.ByName("gender")
+	if database.VerifyAdmin(api.MyToken) {
 
-	NewPrice, _ := strconv.ParseFloat(Price, 64)
-	NewStock, _ := strconv.Atoi(Stock)
-	NewSize, _ := strconv.Atoi(Size)
-	NewRating, _ := strconv.ParseFloat(Rating, 64)
+		Brand := c.Params.ByName("brand")
+		Name := c.Params.ByName("name")
+		Price := c.Params.ByName("price")
+		Stock := c.Params.ByName("stock")
+		Size := c.Params.ByName("size")
+		Rating := c.Params.ByName("rating")
+		Gender := c.Params.ByName("gender")
 
-	products.AddGlasses(Brand, Name, NewPrice, NewStock, NewSize, NewRating, Gender)
+		NewPrice, _ := strconv.ParseFloat(Price, 64)
+		NewStock, _ := strconv.Atoi(Stock)
+		NewSize, _ := strconv.Atoi(Size)
+		NewRating, _ := strconv.ParseFloat(Rating, 64)
 
-	c.JSON(200, gin.H{
-		"message": "Glasses added successfully",
-	})
+		products.AddGlasses(Brand, Name, NewPrice, NewStock, NewSize, NewRating, Gender)
+
+		c.JSON(200, gin.H{
+			"message": "Glasses added successfully",
+		})
+	} else {
+		c.JSON(200, gin.H{
+			"message": "You are not authorized to perform this action",
+		})
+	}
 }
 
 func DeleteGlasses(c *gin.Context) {
-	Name := c.Params.ByName("name")
 
-	products.DeleteGlasses(Name)
-	c.JSON(200, gin.H{
-		"message": "Glasses deleted successfully",
-	})
+	if database.VerifyAdmin(api.MyToken) {
+
+		Name := c.Params.ByName("name")
+
+		products.DeleteGlasses(Name)
+		c.JSON(200, gin.H{
+			"message": "Glasses deleted successfully",
+		})
+	} else {
+		c.JSON(200, gin.H{
+			"message": "You are not authorized to perform this action",
+		})
+	}
 
 }
 
@@ -59,14 +83,22 @@ func ListGlasses(c *gin.Context) {
 }
 
 func BuyGlasses(c *gin.Context) {
-	Name := c.Params.ByName("name")
-	Stock := c.Params.ByName("stock")
 
-	NewStock, _ := strconv.Atoi(Stock)
+	if database.VerifyToken(api.MyToken) {
 
-	products.BuyGlasses(Name, NewStock)
-	c.JSON(200, gin.H{
-		"message": "Glasses bought successfully",
-	})
+		Name := c.Params.ByName("name")
+		Stock := c.Params.ByName("stock")
+
+		NewStock, _ := strconv.Atoi(Stock)
+
+		products.BuyGlasses(Name, NewStock)
+		c.JSON(200, gin.H{
+			"message": "Glasses bought successfully",
+		})
+	} else {
+		c.JSON(200, gin.H{
+			"message": "You are not authorized to perform this action",
+		})
+	}
 
 }
