+++ api/productapi/pantsapi.go
@@ -1,7 +1,9 @@
 package productapi
 
 import (
+	"finalproject/api"
 	"finalproject/products"
+	database "finalproject/user_database"
 	"strconv"
 
 	"github.com/gin-gonic/gin"
@@ -9,39 +11,63 @@ import (
 
 func CreatePantsCategory(c *gin.Context) {
 
-	products.InsertPants()
+	if database.VerifyAdmin(api.MyToken) {
 
-	c.JSON(200, gin.H{
-		"message": "Pants category created successfully",
-	})
+		products.InsertPants()
+
+		c.JSON(200, gin.H{
+			"message": "Pants category created successfully",
+		})
+	} else {
+		c.JSON(200, gin.H{
+			"message": "You are not authorized to perform this action",
+		})
+	}
 }
 
 func AddPants(c *gin.Context) {
-	Brand := c.Params.ByName("brand")
-	Name := c.Params.ByName("name")
-	Price := c.Params.ByName("price")
-	Stock := c.Params.ByName("stock")
-	Size := c.Params.ByName("size")
-	Rating := c.Params.ByName("rating")
-	Gender := c.Params.ByName("gender")
 
-	NewPrice, _ := strconv.ParseFloat(Price, 64)
-	NewStock, _ := strconv.Atoi(Stock)
-	NewSize, _ := strconv.Atoi(Size)
-	NewRating, _ := strconv.ParseFloat(Rating, 64)
+	if database.VerifyAdmin(api.MyToken) {
 
-	products.AddPants(Brand, Name, NewPrice, NewStock, NewSize, NewRating, Gender)
-	c.JSON(200, gin.H{
-		"message": "Pants added successfully",
-	})
+		Brand := c.Params.ByName("brand")
+		Name := c.Params.ByName("name")
+		Price := c.Params.ByName("price")
+		Stock := c.Params.ByName("stock")
+		Size := c.Params.ByName("size")
+		Rating := c.Params.ByName("rating")
+		Gender := c.Params.ByName("gender")
+
+		NewPrice, _ := strconv.ParseFloat(Price, 64)
+		NewStock, _ := strconv.Atoi(Stock)
+		NewSize, _ := strconv.Atoi(Size)
+		NewRating, _ := strconv.ParseFloat(Rating, 64)
+
+		products.AddPants(Brand, Name, NewPrice, NewStock, NewSize, NewRating, Gender)
+		c.JSON(200, gin.H{
+			"message": "Pants added successfully",
+		})
+	} else {
+		c.JSON(200, gin.H{
+			"message": "You are not authorized to perform this action",
+		})
+	}
+
 }
 
 func DeletePants(c *gin.Context) {
-	Name := c.Params.ByName("name")
-	products.DeletePants(Name)
-	c.JSON(200, gin.H{
-		"message": "Pants deleted successfully",
-	})
+
+	if database.VerifyAdmin(api.MyToken) {
+
+		Name := c.Params.ByName("name")
+		products.DeletePants(Name)
+		c.JSON(200, gin.H{
+			"message": "Pants deleted successfully",
+		})
+	} else {
+		c.JSON(200, gin.H{
+			"message": "You are not authorized to perform this action",
+		})
+	}
 
 }
 
@@ -55,14 +81,22 @@ func ListPants(c *gin.Context) {
 }
 
 func BuyPants(c *gin.Context) {
-	Name := c.Params.ByName("name")
-	Stock := c.Params.ByName("stock")
 
-	NewStock, _ := strconv.Atoi(Stock)
+	if database.VerifyToken(api.MyToken) {
 
-	products.BuyPants(Name, NewStock)
-	c.JSON(200, gin.H{
-		"message": "Pants bought successfully",
-	})
+		Name := c.Params.ByName("name")
+		Stock := c.Params.ByName("stock")
+
+		NewStock, _ := strconv.Atoi(Stock)
+
+		products.BuyPants(Name, NewStock)
+		c.JSON(200, gin.H{
+			"message": "Pants bought successfully",
+		})
+	} else {
+		c.JSON(200, gin.H{
+			"message": "You are not authorized to perform this action",
+		})
+	}
 
 }
